requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
prophet==1.1.4
transformers==4.33.2
torch==2.0.1
spacy==3.6.1
nltk==3.8.1
plaid-python==9.1.0
cryptography==41.0.4
bcrypt==4.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
plotly==5.15.0
gunicorn==21.2.0